+++ sql_engine/definition_planner/src/lib.rs
@@ -22,7 +22,6 @@ use query_response::QueryError;
 use storage::{Transaction, TransactionalDatabase};
 use types::SqlType;
 
-#[derive(Clone)]
 pub struct DefinitionPlanner<'p> {
     catalog: CatalogHandler<'p>,
 }
