+++ sql_engine/query_processing/type_check/src/lib.rs
@@ -14,7 +14,6 @@
 
 use data_manipulation_typed_tree::{DynamicTypedTree, StaticTypedTree};
 
-#[derive(Clone)]
 pub struct TypeChecker;
 
 impl TypeChecker {
