+++ sql_engine/query_processing/type_coercion/src/lib.rs
@@ -15,7 +15,6 @@
 use data_manipulation_typed_tree::{DynamicTypedTree, StaticTypedTree};
 use types::SqlType;
 
-#[derive(Clone)]
 pub struct TypeCoercion;
 
 impl TypeCoercion {
